pyproject.toml
[project]
name = "cafe-epos"
version = "0.1.0"
description = "A small part of a café's EPOS system."
authors = [
    {name = "MicahGH", email = "miicah.shiiels@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.6,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py313']

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
    "api/migrations",
    "__init__.py",
]

[tool.lint]
select = [
    "E",
    "F",
    "B",
    "I",
    "D",
]
[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.13.1,<0.14.0)"
]